scraper/new_venv/requirements.txt
requests>=2.30.0
python-dotenv>=1.0.0
supabase>=2.15.0
